requirements.txt
pymysql
mysql-connector-python
werkzeug
flask_api